add bank service

+++ src/main/java/payments/businesslogic/models/Payment.java
@@ -4,13 +4,15 @@
 
 package payments.businesslogic.models;
 
+import java.math.BigDecimal;
+
 public class Payment {
     public String Token;
-    public Double Amount;
+    public BigDecimal Amount;
     public String MerchantId;
     public String Description;
 
-    public Payment(Double amount, String token, String merchantId, String description) {
+    public Payment(BigDecimal amount, String token, String merchantId, String description) {
         Token = token;
         Amount = amount;
         MerchantId = merchantId;
