+++ src/main/java/payments/businesslogic/models/Transaction.java
@@ -4,19 +4,20 @@
 
 package payments.businesslogic.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
 public class Transaction {
     public String Id;
     public String Token;
-    public Double Amount;
+    public BigDecimal Amount;
     public String MerchantId;
     public String CustomerId;
     public String Description;
     public LocalDateTime Time;
 
-    public Transaction(Double amount, String token, String merchantId, String customerId, String description) {
+    public Transaction(BigDecimal amount, String token, String merchantId, String customerId, String description) {
         Id = UUID.randomUUID().toString();
         Token = token;
         Amount = amount;
